pyproject.toml
[project]
name = "nuguard"
version = "0.2.0"
description = "AI application security CLI — SBOM generation, static analysis, cognitive policy validation, and adversarial red-teaming for AI apps"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "NuGuard", email = "oss@nuguard.ai" },
]
keywords = ["ai", "security", "sbom", "red-team", "llm", "prompt-injection", "appsec"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyYAML>=6.0",
    "jsonschema>=4.23",
    "networkx>=3.3",
    "httpx>=0.27",
    "litellm>=1.40",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "asyncpg>=0.29",
    "rich>=13.0",
    "structlog>=25.5.0",
    "cyclonedx-bom>=7.2.2",
    "tree-sitter>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "python-dotenv>=1.0",
    "google-cloud-aiplatform>=1.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/nuguard/nuguard-oss"
Repository = "https://github.com/nuguard/nuguard-oss"
Issues = "https://github.com/nuguard/nuguard-oss/issues"

[project.scripts]
nuguard = "nuguard.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
    "/.claude",
    "/.mypy_cache",
    "/.pytest_cache",
    "/.ruff_cache",
    "/dist",
    "/tmp",
    "/tests",
    "/nuguard/**/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["nuguard"]

[tool.hatch.build.targets.sdist]
include = [
    "/nuguard",
    "/README.md",
    "/pyproject.toml",
    "/Makefile",
    "/canary.example.json",
    "/nuguard.yaml.example",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--import-mode=importlib"
testpaths = ["tests", "nuguard/sbom/tests"]
collect_ignore = [
    "tests/apps/ai-agents-google-adk/test_bedrock_haiku.py",
    "tests/apps/ai-agents-google-adk/test_openai_gpt.py",
]
markers = [
    "redteam_e2e: E2E redteam scan against a live fixture app (opt-in: NUGUARD_REDTEAM_E2E=1)",
    "benchmark_redteam: Full benchmark redteam CLI scan (opt-in: NUGUARD_REDTEAM_BENCHMARK=1)",
    "analyze_e2e: E2E analyze pipeline from GitHub clone (opt-in: NUGUARD_ANALYZE_E2E=1)",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_unused_ignores = false

[dependency-groups]
dev = [
    "boto3>=1.42.91",
    "mypy>=1.19.1",
    "pip-audit>=2.10.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.7",
    "types-pyyaml>=6.0.12.20250915",
]
